requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
email-validator==2.0.0

python-dotenv
